+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - ISPManager recipe and docs.
 - Symfony 5 recipe.
 - Command for checking if a deploy is unlocked. [#2150] [#2150]
+- `runLocally` function now accepts string or a Closure as first parameter. [#2170]
 
 ### Fixed
 - Normalize CRLF to LF new line endings. [#2111]
@@ -17,6 +18,7 @@
 - When symfony_env is set to dev, require-dev are not installed. [#2035]
 - Fixed exit status of rollback command when there are no releases to rollback to. [#2052]
 - When the second parameter $options passed to run() and runLocally(), use it to overwrite default env config. [#2165]
+- Fixed `recipe/check_remote.php` calling `get('bin/php)` in remote host context twice. [#2170]
 
 
 ## v6.8.0
@@ -583,6 +585,7 @@
 - Fixed remove of shared dir on first deploy.
 
 
+[#2170]: https://github.com/deployphp/deployer/issues/2170
 [#2165]: https://github.com/deployphp/deployer/issues/2165
 [#2150]: https://github.com/deployphp/deployer/issues/2150
 [#2111]: https://github.com/deployphp/deployer/pull/2111